+++ src/builtins_util.rs
@@ -1,9 +1,6 @@
 use crate::environment::*;
 use crate::eval::*;
 use crate::types::*;
-use std::borrow::Cow;
-
-use std::convert::{TryFrom, TryInto};
 use std::env;
 
 pub fn param_eval_optional(
@@ -185,293 +182,6 @@ pub fn make_args(
     Ok(list)
 }
 
-pub struct ErrorStrings {}
-
-impl ErrorStrings {
-    pub fn mismatched_type(fn_name: &str, expected: &str, got: &str) -> String {
-        format!("{fn_name}: mismatched type input, expected {expected}, got {got}.")
-    }
-
-    pub fn positive_integer(fn_name: &str, expected: &str, got: &str) -> String {
-        format!("{fn_name}: expected positive {expected}, got {got}.")
-    }
-
-    pub fn positive_integer_no_fn_name(expected: &str, got: &str) -> String {
-        format!("Expected positive {expected}, got {got}.")
-    }
-}
-
-pub trait TryIntoExpression<T>: Sized
-where
-    Self: ToString + TryInto<T>,
-{
-    type Error;
-
-    fn human_readable_dest_type(&self) -> String;
-
-    fn try_into_for(self, fn_name: &str) -> Result<T, LispError> {
-        let hr_src_type = self.to_string();
-        let hr_dest_type = self.human_readable_dest_type();
-        let t = self.try_into();
-        match t {
-            Ok(t) => Ok(t),
-            Err(_) => Err(LispError::new(ErrorStrings::mismatched_type(
-                fn_name,
-                &hr_dest_type,
-                &hr_src_type,
-            ))),
-        }
-    }
-}
-
-impl TryIntoExpression<Expression> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        self.display_type()
-    }
-}
-
-impl From<String> for CharString {
-    fn from(value: String) -> Self {
-        CharString(Cow::from(value))
-    }
-}
-
-impl TryIntoExpression<CharString> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::Char(Cow::from(String::default())).to_string()
-    }
-}
-
-impl From<CharString> for Expression {
-    fn from(src: CharString) -> Self {
-        Expression::alloc_data(ExpEnum::Char(src.0.clone()))
-    }
-}
-
-impl TryFrom<Expression> for CharString {
-    type Error = LispError;
-
-    fn try_from(value: Expression) -> Result<Self, Self::Error> {
-        let val = value.get();
-        match &val.data {
-            ExpEnum::Char(char_string) => Ok(CharString(char_string.to_owned())),
-            _ => Err(LispError::new("Can only convert Char from ExpEnum::Char")),
-        }
-    }
-}
-
-impl TryIntoExpression<String> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::String(Cow::from(String::default()), Default::default()).to_string()
-    }
-}
-
-impl From<String> for Expression {
-    fn from(src: String) -> Self {
-        Expression::alloc_data(ExpEnum::String(src.into(), None))
-    }
-}
-
-impl TryFrom<Expression> for String {
-    type Error = LispError;
-
-    fn try_from(value: Expression) -> Result<Self, Self::Error> {
-        match &value.get().data {
-            ExpEnum::String(cow, _) => Ok(cow.to_string()),
-            ExpEnum::Symbol(sym, _) => Ok(sym.to_string()),
-            ExpEnum::Char(ch) => Ok(ch.to_string()),
-            _ => Err(LispError::new(
-                "Can only convert String from ExpEnum::String.",
-            )),
-        }
-    }
-}
-
-impl From<bool> for Expression {
-    fn from(val: bool) -> Self {
-        if val {
-            Expression::make_true()
-        } else {
-            Expression::make_false()
-        }
-    }
-}
-
-impl TryFrom<Expression> for bool {
-    type Error = LispError;
-    fn try_from(num: Expression) -> Result<Self, Self::Error> {
-        match num.get().data {
-            ExpEnum::True => Ok(true),
-            ExpEnum::False => Ok(false),
-            _ => Err(LispError::new(
-                "Can only convert bool from ExpEnum::True or ExpEnum::False.",
-            )),
-        }
-    }
-}
-
-impl From<()> for Expression {
-    fn from(_: ()) -> Self {
-        Expression::make_nil()
-    }
-}
-
-impl TryIntoExpression<bool> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::True.to_string() + " or " + &ExpEnum::False.to_string()
-    }
-}
-
-impl From<f64> for Expression {
-    fn from(num: f64) -> Self {
-        Expression::alloc_data(ExpEnum::Float(num))
-    }
-}
-
-impl TryFrom<Expression> for f64 {
-    type Error = LispError;
-    fn try_from(num: Expression) -> Result<Self, Self::Error> {
-        match num.get().data {
-            ExpEnum::Float(num) => Ok(num),
-            ExpEnum::Int(num) => Ok(num as f64),
-            _ => Err(LispError::new(
-                "Can only convert f64 from ExpEnum::Float or ExpEnum::Int.",
-            )),
-        }
-    }
-}
-
-impl TryIntoExpression<f64> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::Float(Default::default()).to_string()
-    }
-}
-
-impl From<i64> for Expression {
-    fn from(num: i64) -> Self {
-        Expression::alloc_data(ExpEnum::Int(num))
-    }
-}
-
-impl TryFrom<Expression> for i64 {
-    type Error = LispError;
-    fn try_from(num: Expression) -> Result<Self, Self::Error> {
-        match num.get().data {
-            ExpEnum::Float(num) => Ok(num as i64),
-            ExpEnum::Int(num) => Ok(num),
-            _ => Err(LispError::new(
-                "Can only convert i64 from ExpEnum::Float or ExpEnum::Int.",
-            )),
-        }
-    }
-}
-
-impl TryIntoExpression<i64> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::Int(Default::default()).to_string()
-    }
-}
-
-impl<T> From<Vec<T>> for Expression
-where
-    T: Into<Expression>,
-{
-    fn from(values: Vec<T>) -> Self {
-        let mut vec = vec![];
-        for value in values {
-            vec.push(value.into())
-        }
-        Expression::with_list(vec)
-    }
-}
-
-impl From<usize> for Expression {
-    fn from(num: usize) -> Self {
-        Expression::alloc_data(ExpEnum::Int(num as i64))
-    }
-}
-
-impl TryFrom<Expression> for usize {
-    type Error = LispError;
-    fn try_from(num: Expression) -> Result<Self, Self::Error> {
-        match num.get().data {
-            ExpEnum::Float(num) => {
-                if num >= 0.0 {
-                    Ok(num as usize)
-                } else {
-                    Err(LispError::new(ErrorStrings::positive_integer_no_fn_name(
-                        "Number",
-                        &num.to_string(),
-                    )))
-                }
-            }
-            ExpEnum::Int(num) => {
-                if num >= 0 {
-                    Ok(num as usize)
-                } else {
-                    Err(LispError::new(ErrorStrings::positive_integer_no_fn_name(
-                        "Number",
-                        &num.to_string(),
-                    )))
-                }
-            }
-            _ => Err(LispError::new(
-                "Can only convert usize from positive ExpEnum::Float or ExpEnum::Int.",
-            )),
-        }
-    }
-}
-
-impl TryIntoExpression<usize> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::Int(Default::default()).to_string()
-    }
-}
-
-impl TryFrom<Expression> for (Expression, Expression) {
-    type Error = LispError;
-
-    fn try_from(value: Expression) -> Result<Self, Self::Error> {
-        let data = &value.get().data;
-        let data_type = data.to_string();
-        match data {
-            ExpEnum::Pair(e0, e1) => Ok((e0.clone(), e1.clone())),
-            _ => {
-                let reason = format!("Expected Pair got {}.", data_type);
-                Err(LispError::new(reason))
-            }
-        }
-    }
-}
-
-impl From<(Expression, Expression)> for Expression {
-    fn from((e0, e1): (Expression, Expression)) -> Self {
-        Expression::alloc_data(ExpEnum::Pair(e0, e1))
-    }
-}
-
-impl TryIntoExpression<(Expression, Expression)> for Expression {
-    type Error = LispError;
-
-    fn human_readable_dest_type(&self) -> String {
-        ExpEnum::Pair(Expression::make_nil(), Expression::make_nil()).to_string()
-    }
-}
-
 #[macro_export]
 macro_rules! ret_err_exp_enum {
     ($expression:expr, $(|)? $( $pattern:pat_param )|+ $( if $guard: expr )? $(,)?, $eval:expr, $err:expr) => {
@@ -647,32 +357,4 @@ macro_rules! try_inner_string {
             }
         }
     }};
-}
-
-/// Used by sl_sh_fn macro to embed information at runtime about the parameters of
-/// the rust native function, specifically whether it is a normal Type, or some
-/// supported wrapped type, e.g. Optional.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum TypeHandle {
-    Direct,
-    Optional,
-    VarArgs,
-}
-
-/// Used by sl_sh_fn macro to embed information at runtime about the parameters of
-/// the rust native function, specifically whether it is going to pass the value (a move),
-/// a reference, or mutable reference.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum PassingStyle {
-    Value,
-    Reference,
-    MutReference,
-}
-
-/// Struct used by sl_sh_fn macro to embed information in an array at runtime about each of
-/// the parameters of the rust native function.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct Param {
-    pub handle: TypeHandle,
-    pub passing_style: PassingStyle,
 }
