+++ src/types.rs
@@ -16,10 +16,7 @@ use crate::eval::call_lambda;
 use crate::process::*;
 use crate::symbols::*;
 use crate::unix::fd_to_file;
-use crate::{
-    try_inner_file, try_inner_hash_map, try_inner_hash_map_mut, try_inner_int, try_inner_string,
-    ErrorStrings, LispResult,
-};
+use crate::{ErrorStrings, LispResult, RustProcedureRefMut};
 
 #[derive(Clone, Debug)]
 pub struct LispError {
@@ -1108,133 +1105,11 @@ impl From<&mut ExpEnum> for Expression {
     }
 }
 
-pub struct TypedWrapper<'a, T: ?Sized, U>(&'a U, PhantomData<T>);
+pub struct TypedWrapper<'a, T: ?Sized, U>(pub &'a U, PhantomData<T>);
 
 impl<'a, T: ?Sized, U> TypedWrapper<'a, T, U> {
     pub fn new(src: &'a U) -> TypedWrapper<T, U> {
         TypedWrapper(src, PhantomData::default())
-    }
-}
-
-pub trait RustProcedure<T, F>
-where
-    Self: Sized,
-    F: FnOnce(T) -> LispResult<Expression> + ?Sized,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression>;
-}
-
-pub trait RustProcedureRefMut<T, F>
-where
-    Self: Sized,
-    F: FnOnce(&mut T) -> LispResult<Expression> + ?Sized,
-{
-    fn apply_ref_mut(&mut self, fn_name: &str, fun: F) -> LispResult<Expression>;
-}
-
-pub struct CharString(pub Cow<'static, str>);
-
-pub struct CharStringRef<'a>(pub &'a Cow<'static, str>);
-
-impl<F> RustProcedure<CharString, F> for TypedWrapper<'_, CharString, Expression>
-where
-    F: FnOnce(CharString) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        let got = self.0.display_type();
-        match &self.0.get().data {
-            ExpEnum::Char(str) => {
-                let sym = CharString(str.to_owned());
-                fun(sym)
-            }
-            _ => Err(LispError::new(ErrorStrings::mismatched_type(
-                fn_name,
-                &ExpEnum::Char(Default::default()).to_string(),
-                &got,
-            ))),
-        }
-    }
-}
-
-impl<F> RustProcedure<CharStringRef<'_>, F> for TypedWrapper<'_, CharStringRef<'_>, Expression>
-where
-    F: FnOnce(CharStringRef<'_>) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        let got = self.0.display_type();
-        match &self.0.get().data {
-            ExpEnum::Char(str) => {
-                let sym = CharStringRef(str);
-                fun(sym)
-            }
-            _ => Err(LispError::new(ErrorStrings::mismatched_type(
-                fn_name,
-                &ExpEnum::Char(Default::default()).to_string(),
-                &got,
-            ))),
-        }
-    }
-}
-
-pub struct Symbol(&'static str, SymLoc);
-
-impl<F> RustProcedureRefMut<Symbol, F> for TypedWrapper<'_, Symbol, Expression>
-where
-    F: FnOnce(&mut Symbol) -> LispResult<Expression>,
-{
-    fn apply_ref_mut(&mut self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        let got = self.0.display_type();
-        match &self.0.get().data {
-            ExpEnum::Symbol(str, loc) => {
-                let mut sym = Symbol(str, loc.clone());
-                fun(&mut sym)
-            }
-            _ => Err(LispError::new(ErrorStrings::mismatched_type(
-                fn_name,
-                &ExpEnum::Symbol(Default::default(), SymLoc::None).to_string(),
-                &got,
-            ))),
-        }
-    }
-}
-
-impl<F> RustProcedure<&Symbol, F> for TypedWrapper<'_, Symbol, Expression>
-where
-    F: FnOnce(&Symbol) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        let got = self.0.display_type();
-        match &self.0.get().data {
-            ExpEnum::Symbol(str, loc) => {
-                let sym = Symbol(str, loc.clone());
-                fun(&sym)
-            }
-            _ => Err(LispError::new(ErrorStrings::mismatched_type(
-                fn_name,
-                &ExpEnum::Symbol(Default::default(), SymLoc::None).to_string(),
-                &got,
-            ))),
-        }
-    }
-}
-
-impl<F> RustProcedure<Symbol, F> for TypedWrapper<'_, Symbol, Expression>
-where
-    F: FnOnce(Symbol) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        let got = self.0.display_type();
-        match &self.0.get().data {
-            ExpEnum::Symbol(str, loc) => {
-                let sym = Symbol(str, loc.clone());
-                fun(sym)
-            }
-            _ => Err(LispError::new(ErrorStrings::mismatched_type(
-                fn_name,
-                &ExpEnum::Symbol(Default::default(), SymLoc::None).to_string(),
-                &got,
-            ))),
-        }
     }
 }
 
@@ -1270,155 +1145,6 @@ where
             });
         self.0.data.borrow_mut().data.replace(ExpEnum::HashMap(map));
         x
-    }
-}
-
-impl<F> RustProcedureRefMut<HashMap<&str, Expression>, F>
-    for TypedWrapper<'_, HashMap<&str, Expression>, Expression>
-where
-    F: FnOnce(&mut HashMap<&str, Expression>) -> LispResult<Expression>,
-{
-    fn apply_ref_mut(&mut self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_hash_map_mut!(fn_name, self.0, arg, fun(arg))
-    }
-}
-
-impl<F> RustProcedure<HashMap<&str, Expression>, F>
-    for TypedWrapper<'_, HashMap<&str, Expression>, Expression>
-where
-    F: FnOnce(HashMap<&str, Expression>) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_hash_map!(fn_name, self.0, arg, fun(arg.clone()))
-    }
-}
-
-impl<F> RustProcedure<&str, F> for TypedWrapper<'_, &str, Expression>
-where
-    F: FnOnce(&str) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_string!(fn_name, self.0, arg, fun(arg))
-    }
-}
-
-impl<F> RustProcedure<String, F> for TypedWrapper<'_, String, Expression>
-where
-    F: FnOnce(String) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_string!(fn_name, self.0, arg, {
-            let arg = arg.to_string();
-            fun(arg)
-        })
-    }
-}
-
-impl<F> RustProcedureRefMut<String, F> for TypedWrapper<'_, String, Expression>
-where
-    F: FnOnce(&mut String) -> LispResult<Expression>,
-{
-    fn apply_ref_mut(&mut self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_string!(fn_name, self.0, arg, {
-            let arg = &mut arg.to_string();
-            fun(arg)
-        })
-    }
-}
-
-impl<F> RustProcedure<&Expression, F> for TypedWrapper<'_, Expression, Expression>
-where
-    F: FnOnce(&Expression) -> LispResult<Expression>,
-{
-    fn apply(&self, _fn_name: &str, fun: F) -> LispResult<Expression> {
-        fun(&mut self.0.clone())
-    }
-}
-
-impl<F> RustProcedureRefMut<Expression, F> for TypedWrapper<'_, Expression, Expression>
-where
-    F: FnOnce(&mut Expression) -> LispResult<Expression>,
-{
-    fn apply_ref_mut(&mut self, _fn_name: &str, fun: F) -> LispResult<Expression> {
-        fun(&mut self.0.clone())
-    }
-}
-
-impl<F> RustProcedure<Expression, F> for TypedWrapper<'_, Expression, Expression>
-where
-    F: FnOnce(Expression) -> LispResult<Expression>,
-{
-    fn apply(&self, _fn_name: &str, fun: F) -> LispResult<Expression> {
-        fun(self.0.clone())
-    }
-}
-
-impl<F> RustProcedure<i64, F> for TypedWrapper<'_, i64, Expression>
-where
-    F: FnOnce(i64) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_int!(fn_name, self.0, num, fun(num))
-    }
-}
-
-impl<F> RustProcedure<usize, F> for TypedWrapper<'_, usize, Expression>
-where
-    F: FnOnce(usize) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        match self.0.get().data {
-            ExpEnum::Int(name) => {
-                if name < 0 {
-                    return Err(LispError::new(ErrorStrings::positive_integer(
-                        fn_name,
-                        &ExpEnum::Int(Default::default()).to_string(),
-                        &self.0.to_string(),
-                    )));
-                } else {
-                    fun(name as usize)
-                }
-            }
-            _ => {
-                return Err(LispError::new(ErrorStrings::mismatched_type(
-                    fn_name,
-                    &ExpEnum::Int(Default::default()).to_string(),
-                    &self.0.to_string(),
-                )))
-            }
-        }
-    }
-}
-
-impl<F> RustProcedure<f64, F> for TypedWrapper<'_, f64, Expression>
-where
-    F: FnOnce(f64) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        match &self.0.get().data {
-            ExpEnum::Float(f) => fun(*f),
-            ExpEnum::Int(i) => fun(*i as f64),
-            _ => {
-                let expected = ExpEnum::Float(f64::default()).to_string()
-                    + ", or "
-                    + &ExpEnum::Int(i64::default()).to_string();
-                Err(LispError::new(ErrorStrings::mismatched_type(
-                    fn_name,
-                    &expected,
-                    &self.0.to_string(),
-                )))
-            }
-        }
-    }
-}
-
-impl<F> RustProcedure<Rc<RefCell<FileState>>, F>
-    for TypedWrapper<'_, Rc<RefCell<FileState>>, Expression>
-where
-    F: FnOnce(Rc<RefCell<FileState>>) -> LispResult<Expression>,
-{
-    fn apply(&self, fn_name: &str, fun: F) -> LispResult<Expression> {
-        try_inner_file!(fn_name, self.0, file, fun(file))
     }
 }
